Inline SmartLockBase into SignInDelegate
